% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_confirmations.R
\name{count_confirmations}
\alias{count_confirmations}
\title{Count confirmations or non-confirmations}
\usage{
count_confirmations(filter_confirmation, data_guests)
}
\arguments{
\item{filter_confirmation}{Character vectors. Conditions to be combined with a &}

\item{data_guests}{Tibble.}
}
\value{
Character.
}
\description{
Count confirmations or non-confirmations
}
\examples{
data(data_guests_example)
count_confirmations(filter_confirmation = c("type  == 'Teen'",
                                            "(here_cocktail == 'No' | 
                                              here_diner == 'No' | 
                                              here_sunday == 'No')"),
                    data_guests = data_guests_example)
}
