% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{data_guests_example}
\alias{data_guests_example}
\title{Example for data_guests object}
\format{
4 rows, 40 columns. Columns are name, type, table, announcement, here_cocktail, here_diner, here_sunday, special_diet, menu_diner, time_confirmation
}
\usage{
data("data_guests_example")
}
\description{
This data set is a small example.
}
\examples{
data("data_guests_example")

}
\author{
Margot Brard \email{margot@thinkr.fr}
}
\keyword{datasets}
