% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data.R
\docType{data}
\name{data_guests_table_example}
\alias{data_guests_table_example}
\title{Example for data_guests_table object}
\format{
15 rows, 3 columns. Columns are name, table, here_diner
}
\usage{
data("data_guests_table_example")
}
\description{
This data set is a small example.
}
\examples{
data("data_guests_table_example")

}
\author{
Margot Brard \email{margot@thinkr.fr}
}
\keyword{datasets}
