% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_count_countdown_moments.R
\name{get_count_countdown_moments}
\alias{get_count_countdown_moments}
\title{Construst a count/countdown between two moments}
\usage{
get_count_countdown_moments(
  start_moment,
  end_moment,
  text,
  time_zone = Sys.timezone(),
  language = "en"
)
}
\arguments{
\item{start_moment}{Start moment YYYY-MM-DD HH:MM:SS. Character.}

\item{end_moment}{End moment format YYYY-MM-DD HH:MM:SS. Character.}

\item{text}{Text to print at the end of the count/countdown. Character.}

\item{time_zone}{Time zone. Character. Default is Sys.timezone().}

\item{language}{Language of text. Character. Either "en" or "fr". Default is "en"}
}
\value{
Count/countdown. Character.
}
\description{
Construst a count/countdown between two moments
}
\examples{
get_count_countdown_moments(start_moment = "1998-07-12 21:00:00",
                            end_moment = lubridate::now(tzone = Sys.timezone()),
                            text = "since worldcup 1998")

get_count_countdown_moments(start_moment = "2009-10-14 04:00:00",
                            end_moment = "2021-10-02 15:00:00",
                            text = "between the moment they met and the wedding")
}
