% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_wedding.R
\name{get_map_wedding}
\alias{get_map_wedding}
\title{Create a map for the wedding place}
\usage{
get_map_wedding(data_markers, icon_markers, zoom = 7)
}
\arguments{
\item{data_markers}{Information about the wedding place. Tibble with 3 columns : longitude, latitude, name.}

\item{icon_markers}{Image to be used for the markers. Icon from ion library. Character.}

\item{zoom}{Zoom on the map. Numeric.}
}
\value{
Map.
}
\description{
Create a map for the wedding place
}
\examples{
data_markers <- tibble::tibble(
  longitude = c(2.1203232115689707, 2.119756986412269, 2.120561059665001),
  latitude = c(48.80504860237253, 48.78022930537074, 48.85858486573977),
  name = c(
    "<b>Palace of Versailles</b><br/>Super great place to get married", 
    "<b>A random place to stay</b>", 
    "<b>A random place to stay</b>")
)

get_map_wedding(
  data_markers = data_markers,
  icon_markers = "home"
  )

get_map_wedding(
  data_markers = data_markers,
  icon_markers = "star",
  zoom = 10
)
}
