% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_expenses.R
\name{get_total_expenses}
\alias{get_total_expenses}
\title{Compute the total amount of expenses and convert as character}
\usage{
get_total_expenses(
  .data,
  prefix_money = "$",
  suffix_money = "",
  big_mark_money = ","
)
}
\arguments{
\item{.data}{Tibble.}

\item{prefix_money}{Character.}

\item{suffix_money}{Character.}

\item{big_mark_money}{Character.}
}
\value{
Character.
}
\description{
Compute the total amount of expenses and convert as character
}
\examples{
data(data_expenses_example)
data_expenses_example \%>\% 
  get_total_expenses()
}
