% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{load_wedding_fonts}
\alias{load_wedding_fonts}
\title{Import app fonts.}
\usage{
load_wedding_fonts()
}
\value{
Nothing.
}
\description{
Import app fonts.
}
\examples{
library(ggplot2)

load_wedding_fonts()
extrafont::fonts() # available fonts

ggplot(data = iris) +
  aes(x = Sepal.Width, fill = Species) +
  geom_density(alpha = 0.6, color = NA) +
  labs(title = "Largeur des sépales des iris",
       subtitle = "3 variétés") +
  theme(
    plot.title = element_text(family = "Bauer Bodoni Std"),
    plot.subtitle = element_text(family = "Old Script"),
    text = element_text(family = "Bauer Bodoni Std")
    )
}
