% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_seating_plans.R
\name{plot_all_seating_plans}
\alias{plot_all_seating_plans}
\title{Plot all seating plans}
\usage{
plot_all_seating_plans(
  .data,
  font_family_table = "Times",
  font_family_guests = "Times"
)
}
\arguments{
\item{.data}{Tibble.}

\item{font_family_table}{Character.}

\item{font_family_guests}{Character.}
}
\value{
patchwork object.
}
\description{
Plot all seating plans
}
\examples{
data(data_guests_table_example)
data_guests_table_example \%>\% 
 plot_all_seating_plans()
}
