% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_seating_plan.R
\name{plot_one_seating_plan}
\alias{plot_one_seating_plan}
\title{Plot the seating plan for a given table}
\usage{
plot_one_seating_plan(
  .data,
  id_table,
  font_family_table = "Old Script",
  font_family_guests = "Bauer Bodoni Std"
)
}
\arguments{
\item{.data}{Tibble. Data.}

\item{id_table}{Integer. Id of the table.}

\item{font_family_table}{Character. Font family for the title (number table).}

\item{font_family_guests}{Character. Font family for the guests names.}
}
\value{
A ggplot2 object.
}
\description{
Plot the seating plan for a given table
}
\examples{
data(data_guests_table_example)
data_guests_table_example \%>\% 
  plot_one_seating_plan(id_table = 1)
}
