% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_expenses_to_fr.R
\name{rename_expenses_to_fr}
\alias{rename_expenses_to_fr}
\title{Rename columns of data_expenses to french}
\usage{
rename_expenses_to_fr(.data)
}
\arguments{
\item{.data}{Tibble.}
}
\value{
Tibble.
}
\description{
Rename columns of data_expenses to french
}
\examples{
data(data_expenses_example)
data_expenses_example \%>\% 
  rename_expenses_to_fr()
}
