% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expenses_by_paymaster.R
\name{plot_expenses_by_paymaster}
\alias{plot_expenses_by_paymaster}
\title{Plot expenses by paymaster}
\usage{
plot_expenses_by_paymaster(
  .data,
  prefix_money = "$",
  suffix_money = "",
  big_mark_money = ",",
  main_font_family = "Bauer Bodoni Std",
  second_font_family = "Old Script"
)
}
\arguments{
\item{.data}{Tibble. Data.}

\item{prefix_money}{Character. Money prefix.}

\item{suffix_money}{Character. Money suffix.}

\item{big_mark_money}{Character. Money big mark.}

\item{main_font_family}{Character. Main font family.}

\item{second_font_family}{Character. Second font family (x axis).}
}
\value{
Plot.
}
\description{
Plot expenses by paymaster
}
\examples{
data(data_expenses_example)
data_expenses_example \%>\% 
  plot_expenses_by_paymaster()
}
