% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_seating_plan.R
\name{plot_one_seating_plan}
\alias{plot_one_seating_plan}
\title{Plot the seating plan for a givzn table}
\usage{
plot_one_seating_plan(
  .data,
  id_table,
  font_family_table = "Times",
  font_family_guests = "Times"
)
}
\arguments{
\item{.data}{Tibble.}

\item{id_table}{Integer.}

\item{font_family_table}{Character.}

\item{font_family_guests}{Character.}
}
\value{
A ggplot2 object.
}
\description{
Plot the seating plan for a givzn table
}
\examples{
data(data_guests_table_example)
#data_guests_table_example \%>\% 
  #plot_one_seating_plan(id_table = 1)
}
